requirements.txt

python3-pygame
flake8